ShoppingCart1

+++ QualitySouvenir/QualitySouvenir/Models/Souvenir.cs
@@ -16,7 +16,7 @@ namespace QualitySouvenir.Models
         public string Name { get; set; }
 
         [Required]
-        public Double Price { get; set; }
+        public decimal Price { get; set; }
 
         public string Description { get; set; }
